updating plot pt pseudo section
